requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
groq==1.1.2
mempalace==3.0.0
chromadb>=0.4.0
psycopg2-binary==2.9.11
pydantic-settings==2.5.2
python-dotenv==1.0.1
